@@ -10,16 +10,16 @@
     }
   ],
   "require": {
-    "php": "^5.6 || ^7.0",
+    "php": "^7.1",
     "psr/http-message": "^1.0",
     "container-interop/container-interop": "^1.1",
     "zendframework/zend-expressive-router": "^1.0",
     "zendframework/zend-servicemanager": "^3.0",
 
-    "dotkernel/dot-rbac": "0.6.x-dev",
-    "dotkernel/dot-authorization": "0.6.x-dev",
-    "dotkernel/dot-event": "0.6.x-dev",
-    "dotkernel/dot-helpers": "0.6.x-dev"
+    "dotkernel/dot-rbac": "dev-develop as 0.6.x-dev",
+    "dotkernel/dot-authorization": "dev-develop as 0.6.x-dev",
+    "dotkernel/dot-event": "dev-develop as 0.6.x-dev",
+    "dotkernel/dot-helpers": "dev-develop as 0.6.x-dev"
   },
   "require-dev": {
     "phpunit/phpunit": "^4.8",
@@ -28,9 +28,9 @@
     "zendframework/zend-diactoros": "^1.3",
     "zendframework/zend-stdlib": "^3.0",
 
-    "dotkernel/dot-controller": "0.6.x-dev",
-    "dotkernel/dot-authentication": "0.6.x-dev",
-    "dotkernel/dot-flashmessenger": "0.6.x-dev"
+    "dotkernel/dot-controller": "dev-develop as 0.6.x-dev",
+    "dotkernel/dot-authentication": "dev-develop as 0.6.x-dev",
+    "dotkernel/dot-flashmessenger": "dev-develop as 0.6.x-dev"
   },
   "autoload": {
     "psr-4": {
