@@ -7,6 +7,8 @@
  * Time: 12:50 AM
  */
 
+declare(strict_types=1);
+
 namespace Dot\Rbac\Guard\Factory;
 
 use Dot\Rbac\Guard\Options\RbacGuardOptions;
@@ -20,10 +22,11 @@ class RbacGuardOptionsFactory
 {
     /**
      * @param ContainerInterface $container
+     * @param $requestedName
      * @return RbacGuardOptions
      */
-    public function __invoke(ContainerInterface $container)
+    public function __invoke(ContainerInterface $container, $requestedName)
     {
-        return new RbacGuardOptions($container->get('config')['dot_authorization']);
+        return new $requestedName($container->get('config')['dot_authorization']);
     }
 }
